Fix E2E tests (skeleton for Monitoring Alerts card)
