+++ packages/yourloops/test/integration/pages/patient-dashboard/patient-dashboard.hcp.spec.tsx
@@ -30,7 +30,7 @@ import { logoutMock, mockAuth0Hook } from '../../mock/auth0.hook.mock'
 import { buildAvailableTeams, mockTeamAPI, myThirdTeamId, myThirdTeamName } from '../../mock/team.api.mock'
 import { completeDashboardData, mockDataAPI, twoWeeksOldDashboardData } from '../../mock/data.api.mock'
 import { mockNotificationAPI } from '../../mock/notification.api.mock'
-import { patient1Id, patient1Info, patientWithMmolId } from '../../data/patient.api.data'
+import { patient1Id, patient1Info, patient1Metrics, patientWithMmolId } from '../../data/patient.api.data'
 import { mockChatAPI } from '../../mock/chat.api.mock'
 import { mockMedicalFilesAPI, mockMedicalFilesApiEmptyResult } from '../../mock/medical-files.api.mock'
 import { mockDirectShareApi } from '../../mock/direct-share.api.mock'
@@ -176,6 +176,7 @@ describe('Patient dashboard for HCP', () => {
       ...patient1Info,
       invitationStatus: UserInviteStatus.Accepted
     }])
+    jest.spyOn(PatientApi, 'getPatientsMetricsForHcp').mockResolvedValue([patient1Metrics])
 
     const appMainLayoutParams: AppMainLayoutHcpParams = {
       footerHasLanguageSelector: false,
