+++ packages/yourloops/test/unit/lib/patient/patients.hook.test.tsx
@@ -96,6 +96,7 @@ describe('Patients hook', () => {
         }
       }
     })
+    jest.spyOn(PatientApi, 'getPatientsMetricsForHcp').mockResolvedValue([])
   })
 
   async function renderPatientsHook(patients: Patient[]) {
